refactor source data
