+++ frontend/src/views/WheresMyGeneV2/components/InfoPanel/components/SourceData/index.tsx
@@ -1,36 +1,12 @@
 import { List, ListItem } from "@czi-sds/components";
-import { useContext, useMemo } from "react";
 import { track } from "src/common/analytics";
 import { EVENTS } from "src/common/analytics/events";
-import {
-  aggregateCollectionsFromDatasets,
-  useFilterDimensions,
-} from "src/common/queries/wheresMyGene";
-import { StateContext } from "src/views/WheresMyGeneV2/common/store";
 import { Content, InfoText, ListSubheader, Wrapper } from "./style";
 import { ROUTES } from "src/common/constants/routes";
-
-interface Collection {
-  name: string;
-  url: string;
-  datasets: { id: string; label: string }[];
-}
-
-interface Collections {
-  [name: string]: Collection;
-}
+import { useConnect } from "./connect";
 
 export default function SourceData(): JSX.Element {
-  const { selectedFilters } = useContext(StateContext);
-  const { data: filterDimensions } = useFilterDimensions();
-  let { datasets = [] } = filterDimensions;
-  if (selectedFilters.datasets.length > 0)
-    datasets = datasets.filter((dataset) =>
-      selectedFilters.datasets.includes(dataset.id)
-    );
-  const collections: Collections = useMemo(() => {
-    return aggregateCollectionsFromDatasets(datasets);
-  }, [datasets]);
+  const { collections } = useConnect();
 
   return (
     <Wrapper>
