+++ frontend/src/views/WheresMyGeneV2/components/InfoPanel/components/SourceData/types.ts
@@ -1,0 +1,9 @@
+export interface Collection {
+  name: string;
+  url: string;
+  datasets: { id: string; label: string }[];
+}
+
+export interface Collections {
+  [name: string]: Collection;
+}
